Interceptor can destroy incoming missile
When the range values of the missile and interceptor converge, the missile
is considered hit and the game reports it as destroyed.

At this point it's time for some refactoring to make the code easier to
work with. The missile and interceptor are worth making into a struct
by now.